Dev a main

+++ src/pages/Carrito.jsx
@@ -1,13 +1,17 @@
 import "./Carrito.css";
-import { useState, useEffect } from "react";
+import { useState, useEffect, useContext } from "react";
 import Formcarrito from "../components/Formcarrito";
 import { Button, Modal } from "react-bootstrap";
 import Envio from "../components/EnvioForm";
 import Swal from "sweetalert2";
+import { Link, useNavigate } from "react-router-dom";
+import { CartContext } from "../context/ContextProvider";
 
-function Carrito({ setCartCount }) {
+function Carrito() {
+  const { setCartCount } = useContext(CartContext);
   const [showModal, setShowModal] = useState(false);
   const [cartItems, setCartItems] = useState([]);
+  const navigate = useNavigate();
 
   useEffect(() => {
     const cartItems = JSON.parse(localStorage.getItem("cartItems")) || [];
@@ -45,34 +49,71 @@ function Carrito({ setCartCount }) {
   };
 
   const handleRemoveFromCart = (producto) => {
-    const newCartItems = cartItems.filter((item) => item._id !== producto._id);
-    if (newCartItems.length !== cartItems.length) {
-      setCartItems(newCartItems);
-      localStorage.setItem("cartItems", JSON.stringify(newCartItems));
-      const count = newCartItems.length;
-      setCartCount(count);
-    }
+    Swal.fire({
+      title: "Eliminar producto",
+      text: `¿Estás seguro que deseas eliminar ${producto.name} del carrito?`,
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonText: "Eliminar",
+      cancelButtonText: "Cancelar",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        const newCartItems = cartItems.filter(
+          (item) => item._id !== producto._id
+        );
+        if (newCartItems.length !== cartItems.length) {
+          setCartItems(newCartItems);
+          localStorage.setItem("cartItems", JSON.stringify(newCartItems));
+          const count = newCartItems.reduce(
+            (count, item) => count + item.quantity,
+            0
+          );
+          setCartCount(count);
+        }
+      }
+    });
   };
 
   const handlePayment = () => {
     const valKey = localStorage.getItem("validation");
     const valKeyTwo = localStorage.getItem("validation2");
-    if (valKey && valKeyTwo) {
-      Swal.fire({
-        position: "top-center",
-        icon: "success",
-        title:
-          "¡Compra realizada con exito, se enviara su factura al correo electronico!",
-        showConfirmButton: false,
-        timer: 1500,
-      });
+    const token = localStorage.getItem("token");
 
-      handleCloseModal();
-      setCartItems([]);
-      localStorage.removeItem("cartItems");
-      localStorage.removeItem("validation");
-      localStorage.removeItem("validation2");
-      window.location.reload();
+    if (valKey === "true" && valKeyTwo === "true") {
+      if (token) {
+        Swal.fire({
+          position: "center",
+          icon: "success",
+          title:
+            "¡Compra realizada con éxito, se enviará su factura al correo electrónico!",
+          showConfirmButton: false,
+          timer: 2500,
+        });
+
+        handleCloseModal();
+        setCartCount(0);
+        setCartItems([]);
+        localStorage.removeItem("cartItems");
+        localStorage.removeItem("validation");
+        localStorage.removeItem("validation2");
+        navigate("/");
+      } else {
+        Swal.fire({
+          position: "center",
+          icon: "info",
+          title: "Debe iniciar sesión para realizar la compra",
+          showConfirmButton: false,
+          timer: 2500,
+        });
+      }
+    } else {
+      Swal.fire({
+        position: "center",
+        icon: "info",
+        title: "Verifique los datos ingresados",
+        showConfirmButton: false,
+        timer: 2500,
+      });
     }
   };
 
@@ -105,7 +146,12 @@ function Carrito({ setCartCount }) {
                         Eliminar
                       </button>
                       <button className="btn btn-primary">
-                        Ver más productos
+                        <Link
+                          to="/productos"
+                          className="text-light text-decoration-none"
+                        >
+                          Ver más productos
+                        </Link>
                       </button>
                       <Button variant="warning" onClick={handleShowModal}>
                         Comprar Ahora ({item.quantity})
